requirements.txt
# Core dependencies
numpy>=1.19.0
pandas>=1.2.0
PyYAML>=5.4.1

# Machine Learning & AI
tensorflow>=2.8.0
scikit-learn>=0.24.0

# Technical Analysis
ta>=0.7.0  # Technical Analysis library

# Exchange Integration
ccxt>=2.0.0  # Cryptocurrency exchange trading library
websockets>=10.0  # For real-time data streaming

# Database
SQLAlchemy>=1.4.0  # SQL toolkit and ORM

# API & Web
requests>=2.26.0
aiohttp>=3.8.0  # Async HTTP client/server
python-dotenv>=0.19.0  # Environment variable management

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3

# Development & Testing
pytest>=6.2.5
black>=21.12b0  # Code formatter
mypy>=0.910  # Static type checker 